pyproject.toml
[tool.poetry]
name = "agent-mode"
version = "0.1.0"
description = "AI Agent System Core"
authors = ["NxPKG <116948796+NxPKG@users.noreply.github.com>"]
readme = "README.md"
packages = [
  { include = "agent" },
  { include = "agent.plugins" }
]

[tool.poetry.dependencies]
python = "^3.9"
openai = ">=1.77.0,<2.0.0"
rich = ">=14.0.0,<15.0.0"
boto3 = ">=1.38.9,<2.0.0"
kubernetes = ">=32.0.1,<33.0.0"
pygithub = ">=2.6.1,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
typer = ">=0.15.3,<0.16.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
